notebooks/requirements.txt
# Jupyter Notebook Requirements for HR Resume Search API
# Python 3.12+ compatible versions

# Core Dependencies
jupyter>=1.0.0
notebook>=7.0.0
ipykernel>=6.25.0
ipywidgets>=8.1.0

# Data Science & Visualization
numpy>=1.24.0
pandas>=2.0.0
matplotlib>=3.7.0
seaborn>=0.12.0
plotly>=5.17.0
scipy>=1.11.0

# HTTP & Async
httpx>=0.25.0
aiofiles>=23.0.0
asyncio-mqtt>=0.16.0

# MCP Client
mcp>=0.1.0

# Database
sqlalchemy>=2.0.0
asyncpg>=0.29.0
psycopg2-binary>=2.9.0

# Environment & Config
python-dotenv>=1.0.0
pydantic>=2.4.0
pydantic-settings>=2.0.0

# Performance & Monitoring
prometheus-client>=0.18.0
psutil>=5.9.0

# Rich Terminal Output (optional but recommended)
rich>=13.5.0

# Testing & Quality
pytest>=7.4.0
pytest-asyncio>=0.21.0

# Utilities
python-dateutil>=2.8.0
pytz>=2023.3